requirements.txt
amqp==5.1.0
APScheduler==3.6.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==4.2.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
Django==4.0.3
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
mysqlclient==2.1.0
prompt-toolkit==3.0.28
python-crontab==2.6.0
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
